Snapshotter cleanup fix
During snapshotter cleanup lcow/wcow snapshotters try to rename the snapshot directory
before cleaning it up. This rename operation sometimes fails if the sandbox.vhdx file
inside that directory is still open in some process (ideally it should not be open but can
stay open in case of a crash or exception) or if the directory is busy for some other
reason. However, if this rename operation fails we return that error and then the entire
snapshot garbage collection operation fails. Due to this we end up not cleaning other
snapshots which otherwise could have been cleaned up without any errors. If this goes on
for a long time we will end up filling the entire disk with stale snapshots. This change
fixes that issue by continuing the snapshot cleanup if the rename operation fails because
of such open handles.

Signed-off-by: Amit Barve <ambarve@microsoft.com>